Updated special characters list
Two changes based on password policies we've encountered, granted, if we try to convene on all policies we'll eventually hit a spot where some services won't accept any generated passwords, so I'm only doing this as it is frequently encountered.

First is the set of symbols is too wide, used Steve Gibson's short symbol list:
!#%+:=?@

Wrote a lua script based on leetspeak.lua to swap the characters and once it was working I noticed it's a feature under advanced tab, damn it.

The other issue is at least one site wanted a second capital letter, it wanted two of each type, which seems excessively strict, but not a big entropy hit, so complied by adding a single capital letter to the end of the password.

At that point the only other problematic requirement was that the length was too long for some sites, but I've yet to hear a compelling reason for shortening a password when it's ultimately hashed. It could be an abundance of caution for the chance of some sort of SQL injection.

Most passwords are landing in the 20-30 length range, I could script powershell or lua to run it again if it detects a password too long but that adds additional processing time and delay in getting a password, so I'd rather avoid that, the user can just run it again and produce a second hopefully shorter password for these rare cases, that's better than reducing entropy by dropping a number of word combinations.